refactor: DTO에 validation 어노테이션 설정
