+++ src/main/java/kitchenpos/ui/dto/menugroup/MenuGroupRequest.java
@@ -1,9 +1,12 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.menugroup;
+
+import javax.validation.constraints.NotEmpty;
 
 import kitchenpos.domain.MenuGroup;
 
 public class MenuGroupRequest {
 
+    @NotEmpty
     private String name;
 
     protected MenuGroupRequest() {
