+++ src/main/java/kitchenpos/ui/dto/product/ProductRequest.java
@@ -1,12 +1,20 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.product;
 
 import java.math.BigDecimal;
+
+import javax.validation.constraints.DecimalMin;
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 
 import kitchenpos.domain.Product;
 
 public class ProductRequest {
 
+    @NotBlank
     private String name;
+
+    @NotNull
+    @DecimalMin("0")
     private BigDecimal price;
 
     protected ProductRequest() {
