+++ src/main/java/kitchenpos/ui/dto/product/ProductResponses.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.product;
 
 import java.util.Collections;
 import java.util.List;
